Replace reference RT allocator.
