+++ .gitmodules
@@ -44,3 +44,6 @@
 	path = src/ports/memmap2-rs
 	url = ../../twizzler-operating-system/memmap2-rs.git
 	branch = twizzler
+[submodule "src/ports/ferroc"]
+	path = src/ports/ferroc
+	url = git@github.com:dbittman/ferroc
