+++ src/lib/naming/naming-core/src/store.rs
@@ -2,6 +2,7 @@ use core::str;
 use std::{
     path::{Component, Path, PathBuf},
     sync::Arc,
+    time::Instant,
 };
 
 use bitflags::bitflags;
