+++ src/rt/monitor/Cargo.toml
@@ -21,6 +21,7 @@ talc = "4.4"
 happylock = "0.5.0"
 parking_lot = "*"
 naming-core = { path = "../../lib/naming/naming-core/" }
+backtracer_core = "*"
 
 [features]
 secgate-impl = []
